+++ src/main/java/roomescape/reservation/Reservation.java
@@ -1,6 +1,7 @@
-package roomescape.domain;
+package roomescape.reservation;
 
 import java.time.LocalDate;
+import roomescape.reservationtime.ReservationTime;
 
 public class Reservation {
