+++ src/main/java/roomescape/reservationtime/dto/request/ReservationTimeRequest.java
@@ -1,8 +1,8 @@
-package roomescape.dto.request;
+package roomescape.reservationtime.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.reservationtime.ReservationTime;
 
 public record ReservationTimeRequest(
         @NotNull LocalTime startAt
