+++ src/main/java/roomescape/ui/RoomescapeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
