+++ src/test/java/roomescape/SpringJdbcTest.java
@@ -13,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.dto.response.ReservationResponse;
+import roomescape.reservation.dto.response.ReservationResponse;
 
 import org.junit.jupiter.api.Test;
